Break reducers test into separate suite per reducer
